Modification update
Modify the functionality of urlComposer
