+++ .npmignore
@@ -1,0 +1,6 @@
+test
+.github
+notes
+node_modules
+example
+dist/web/*
