+++ src/lib/queryString.js
@@ -1,5 +1,45 @@
-const {each, getTypeof, indexOf} = require("structkit");
+const {each, varExtend, getTypeof, indexOfNotExist, indexOf} = require("structkit");
+const {configQueryString} = require("./config");
 const zero =0;
+
+
+/**
+ * Query String stringify
+ *
+ * @since 1.0.0
+ * @category Seq
+ * @param {any} value Passing object to convert string
+ * @param {any=} config Conversion delimeter
+ * @returns {any} Returns the total.
+ * @example
+ *
+ * qsStringify({"test": 11,"test2": 11})
+ *=> test=1&test2=11
+ */
+function qsStringify (value, config) {
+
+    if (indexOfNotExist([
+        "json",
+        "array"
+    ], getTypeof(value))) {
+
+        return "";
+
+    }
+
+    const referenceValue = [];
+    const defaultConfig = varExtend(configQueryString, config);
+
+    each(value, function (key, val) {
+
+        parseStringConvert(key, val, getTypeof(val), defaultConfig, referenceValue);
+
+    });
+
+    return referenceValue.join(defaultConfig.newLineSeparator);
+
+}
+
 
 /**
  * Parse query string to object
@@ -45,4 +85,4 @@ const parseStringConvert=function (key, value, type, config, reference) {
 
 };
 
-exports.parseStringConvert = parseStringConvert;
+exports.qsStringify = qsStringify;
